[Translations] Consolidate Translation models & remove deprecated classes (#7808)
* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

* [Translations] Consolidate Translation models & remove deprecated classes #7504

+++ bundles/AdminBundle/Controller/Admin/TranslationController.php
@@ -74,10 +74,11 @@ class TranslationController extends AdminController
         $overwrite = $merge ? false : true;
 
         if ($admin) {
-            $delta = Translation\Admin::importTranslationsFromFile($tmpFile, $overwrite, Tool\Admin::getLanguages(), $dialect);
+            $delta = Translation::importTranslationsFromFile($tmpFile, Translation::DOMAIN_ADMIN, $overwrite, Tool\Admin::getLanguages(), $dialect);
         } else {
-            $delta = Translation\Website::importTranslationsFromFile(
+            $delta = Translation::importTranslationsFromFile(
                 $tmpFile,
+                Translation::DOMAIN_DEFAULT,
                 $overwrite,
                 $this->getAdminUser()->getAllowedLanguagesForEditingWebsiteTranslations(),
                 $dialect
@@ -159,22 +160,19 @@ class TranslationController extends AdminController
         $admin = $request->get('admin');
         $this->checkPermission(($admin ? 'admin_' : '') . 'translations');
 
+        $domain = Translation::DOMAIN_DEFAULT;
         if ($admin) {
-            $class = '\\Pimcore\\Model\\Translation\\Admin';
-        } else {
-            $class = '\\Pimcore\\Model\\Translation\\Website';
+            $domain = Translation::DOMAIN_ADMIN;
         }
-
-        $tableName = call_user_func($class . '\\Dao::getTableName');
+        $translation = new Translation();
+        $translation->setDomain($domain);
+        $tableName = $translation->getDao()->getDatabaseTableName();
 
         // clear translation cache
         Translation::clearDependentCache();
 
-        if ($admin) {
-            $list = new Translation\Admin\Listing();
-        } else {
-            $list = new Translation\Website\Listing();
-        }
+        $list = new Translation\Listing();
+        $list->setDomain($domain);
 
         $joins = [];
 
@@ -201,10 +199,11 @@ class TranslationController extends AdminController
         // fill with one dummy translation if the store is empty
         if (empty($translationObjects)) {
             if ($admin) {
-                $t = new Translation\Admin();
+                $t = new Translation();
+                $t->setDomain(Translation::DOMAIN_ADMIN);
                 $languages = Tool\Admin::getLanguages();
             } else {
-                $t = new Translation\Website();
+                $t = new Translation();
                 $languages = $this->getAdminUser()->getAllowedLanguagesForViewingWebsiteTranslations();
             }
 
@@ -272,11 +271,10 @@ class TranslationController extends AdminController
             $csv .= implode(';', $tempRow) . "\r\n";
         }
 
-        $suffix = $admin ? 'admin' : 'website';
         $response = new Response("\xEF\xBB\xBF" . $csv);
         $response->headers->set('Content-Encoding', 'UTF-8');
         $response->headers->set('Content-Type', 'text/csv; charset=UTF-8');
-        $response->headers->set('Content-Disposition', 'attachment; filename="export_ ' . $suffix . '_translations.csv"');
+        $response->headers->set('Content-Disposition', 'attachment; filename="export_ ' . $domain . '_translations.csv"');
         ini_set('display_errors', false); //to prevent warning messages in csv
 
         return $response;
@@ -299,13 +297,13 @@ class TranslationController extends AdminController
                 $t = null; // reset
 
                 try {
-                    $t = Translation\Admin::getByKey($translationData);
+                    $t = Translation::getByKey($translationData, Translation::DOMAIN_ADMIN);
                 } catch (\Exception $e) {
                     Logger::log($e);
                 }
-                if (!$t instanceof Translation\Admin) {
-                    $t = new Translation\Admin();
-
+                if (!$t instanceof Translation) {
+                    $t = new Translation();
+                    $t->setDomain(Translation::DOMAIN_ADMIN);
                     $t->setKey($translationData);
                     $t->setCreationDate(time());
                     $t->setModificationDate(time());
@@ -340,27 +338,30 @@ class TranslationController extends AdminController
 
         $this->checkPermission(($admin ? 'admin_' : '') . 'translations');
 
+        $domain = Translation::DOMAIN_DEFAULT;
         if ($admin) {
-            $class = '\\Pimcore\\Model\\Translation\\Admin';
-        } else {
-            $class = '\\Pimcore\\Model\\Translation\\Website';
+            $domain = Translation::DOMAIN_ADMIN;
         }
 
-        $tableName = call_user_func($class . '\\Dao::getTableName');
+        $translation = new Translation();
+        $translation->setDomain($domain);
+        $tableName = $translation->getDao()->getDatabaseTableName();
 
         // clear translation cache
-        Translation\Website::clearDependentCache();
+        Translation::clearDependentCache();
 
         if ($request->get('data')) {
             $data = $this->decodeJson($request->get('data'));
 
             if ($request->get('xaction') == 'destroy') {
-                $t = $class::getByKey($data['key']);
-                $t->delete();
+                $t = Translation::getByKey($data['key'], $domain);
+                if ($t instanceof Translation) {
+                    $t->delete();
+                }
 
                 return $this->adminJson(['success' => true, 'data' => []]);
             } elseif ($request->get('xaction') == 'update') {
-                $t = $class::getByKey($data['key']);
+                $t = Translation::getByKey($data['key'], $domain);
 
                 foreach ($data as $key => $value) {
                     $key = preg_replace('/^_/', '', $key, 1);
@@ -384,12 +385,13 @@ class TranslationController extends AdminController
 
                 return $this->adminJson(['data' => $return, 'success' => true]);
             } elseif ($request->get('xaction') == 'create') {
-                $t = $class::getByKey($data['key']);
+                $t = Translation::getByKey($data['key'], $domain);
                 if ($t) {
-                    throw new \Exception($translator->trans('identifier_already_exists', [], 'admin'));
+                    throw new \Exception($translator->trans('identifier_already_exists', [], $domain));
                 }
 
-                $t = new $class();
+                $t = new Translation();
+                $t->setDomain($domain);
                 $t->setKey($data['key']);
                 $t->setCreationDate(time());
                 $t->setModificationDate(time());
@@ -412,11 +414,8 @@ class TranslationController extends AdminController
             }
         } else {
             // get list of types
-            if ($admin) {
-                $list = new Translation\Admin\Listing();
-            } else {
-                $list = new Translation\Website\Listing();
-            }
+            $list = new Translation\Listing();
+            $list->setDomain($domain);
 
             $validLanguages = $admin ? Tool\Admin::getLanguages() : $this->getAdminUser()->getAllowedLanguagesForViewingWebsiteTranslations();
 
@@ -498,7 +497,7 @@ class TranslationController extends AdminController
 
     /**
      * @param array $joins
-     * @param Translation\AbstractTranslation\Listing $list
+     * @param Translation\Listing $list
      * @param string $tableName
      * @param array $filters
      */
@@ -650,9 +649,17 @@ class TranslationController extends AdminController
      */
     public function cleanupAction(Request $request)
     {
-        $listClass = '\\Pimcore\\Model\\Translation\\' . ucfirst($request->get('type')) . '\\Listing';
+        $admin = $request->get('admin');
+
+        $listClass = '\\Pimcore\\Model\\Translation\\Listing';
+        $domain = Translation::DOMAIN_DEFAULT;
+        if ($admin) {
+            $domain = Translation::DOMAIN_ADMIN;
+        }
+
         if (Tool::classExists($listClass)) {
             $list = new $listClass();
+            $list->setDomain($domain);
             $list->cleanup();
 
             \Pimcore\Cache::clearTags(['translator', 'translate']);
@@ -1219,9 +1226,12 @@ class TranslationController extends AdminController
 
         $dataList = json_decode($request->get('data'), true);
 
-        $classname = '\\Pimcore\\Model\\Translation\\' . ucfirst($translationType);
+        $domain = Translation::DOMAIN_DEFAULT;
+        if ($translationType == "admin") {
+            $domain = Translation::DOMAIN_ADMIN;
+        }
         foreach ($dataList as $data) {
-            $t = $classname::getByKey($data['key'], true);
+            $t = Translation::getByKey($data['key'], $domain, true);
             $newValue = htmlspecialchars_decode($data['current']);
             $t->addTranslation($data['lg'], $newValue);
             $t->setModificationDate(time());
